* os_Windows/os_internal.h: rename instead 'handle' a more unique identifier 'winHandleMutex' and 'winHandleWaitNotify', better for data viewing while debug.   * ObjectJc, ReflectionBaseTypesJc.c, Reflection_AllJc.c: The reflections of ObjectJc and some derived classes are containes in ObjectJc now. It contains the overrideable methods toString() etc. which are need for Linking. This reflections are not generated automaticly up to now, it is constant.   * ReflectionJc.c: The check whether a pointer to Reflection is kREFLECTION_LastConstant needs a 64-bit-integer comparison on a 64-bit System.   * FBaccess_Inspc.c: improved   * InterProcessCommRxExec_Ipc.stc: This is a module which contains an receiver with its thread with a InterProcessComm interface.

+++ source/Jc/ReflectionJc.c
@@ -235,7 +235,7 @@ FieldJcArray* new_FieldJcArray(int size)
 
 
 const char* getName_ClassJc(ClassJc const* ythis)
-{ if( ((int)(ythis)) < kREFLECTION_LastConstant )
+{ if( ((uint64)(ythis)) < kREFLECTION_LastConstant )
   { ythis = simpleTypes[(int)(ythis)];
   }
   return ythis->name;
@@ -254,7 +254,7 @@ METHOD_C bool isPrimitive_ClassJc(ClassJc const* ythis)
 
 METHOD_C int32 getModifiers_ClassJc(ClassJc const* ythis)
 {
-  if( ((int)(ythis)) < kREFLECTION_LastConstant )
+  if( ((uint64)(ythis)) < kREFLECTION_LastConstant )
   { ythis = simpleTypes[(int)(ythis)];
   }
   return ythis->modifiers;
@@ -401,7 +401,7 @@ int32 invoke_MethodJcSi(const MethodJc* ythis, MemSegmJc objP, int32ARRAY* param
  */
 ClassJc const* getType_FieldJc(FieldJc const* ythis)
 { ClassJc const* type = ythis->type_;
-  if( ((int)(type)) < kREFLECTION_LastConstant )
+  if( ((uint64)(type)) < kREFLECTION_LastConstant )
   { type = simpleTypes[(int)(type)];
   }
   return type;
